Синхронизация с гит

+++ views/refereecomp/index.php
@@ -1,5 +1,5 @@
 <h2>Судьи</h2>
-<?if ($_SESSION['userType'] == 3) {?>
+<?if (($_SESSION['userType'] == 3) || ($_SESSION['userComp'][$_GET['comp']] == 1)) {?>
     <div class="main-fieldWrapper">
         <button class="referee-compAddBtn main-btn roster-submit">Добавить</button>
     </div>
@@ -23,13 +23,45 @@
         </div>
     </form>
 <?}?>
-<div class="listview">
+<table class="datagrid roster-view datagrid_zebra">
+    <colgroup>
+        <col width="60px"/>
+        <col/>
+        <col width="80px"/>
+        <col width="80px"/>
+        <?if ($_SESSION['userType'] == 3) {?>
+            <col width="50px"/>
+        <?}?>
+    </colgroup>
+    <tbody class="datagrid_tbody">
     <?php for ($i = 0; $i < count($answer['referee']); $i++) {?>
-        <div class="listview-item">
-            <a target="_blank" href="/?r=person/view&person=<?=$answer['referee'][$i]['id']?>"><?=implode(' ', array($answer['referee'][$i]['surname'], $answer['referee'][$i]['name'], $answer['referee'][$i]['patronymic']))?></a>
-            судейство <?=$answer['referee'][$i]['exp'] ? 'с '.$answer['referee'][$i]['exp'] : 'без опыта'?>,
-            игра <?=$answer['referee'][$i]['expplay'] ? 'с '.$answer['referee'][$i]['expplay'] : 'без опыта'?>
-            <a class="main-dellink" href="/?r=refereecomp/delete&rc=<?=$answer['referee'][$i]['rc']?>&comp=<?=$_GET['comp']?>">[X]</a>
-        </div>
+        <tr>
+            <td>
+                <?if ($answer['referee'][$i]['avatar']) {?>
+                    <a class="referee-clAvatar" target="_blank" href="/?r=person/view&person=<?=$answer['referee'][$i]['id']?><?=$filter?>">
+                        <img style="width:50px" src="//<?=$HOST?>/upload/<?=$answer['referee'][$i]['avatar']?>">
+                    </a>
+                <?} else {?>
+                    <a class="referee-clAvatar" target="_blank" href="/?r=person/view&person=<?=$answer['referee'][$i]['id']?><?=$filter?>">
+                        <div class="main-noPhoto">?</div>
+                    </a>
+                <?}?>
+            </td>
+            <td>
+                <a target="_blank" href="/?r=person/view&person=<?=$answer['referee'][$i]['id']?>"><?=implode(' ', array($answer['referee'][$i]['surname'], $answer['referee'][$i]['name'], $answer['referee'][$i]['patronymic']))?></a>
+            </td>
+            <td>
+                <?=common_dateFromSQL($answer['referee'][$i]['birthdate'])?>
+            </td>
+            <td>
+                <?=$answer['referee'][$i]['country']?>
+            </td>
+            <?if (($_SESSION['userType'] == 3) || ($_SESSION['userComp'][$_GET['comp']] == 1)) {?>
+                <td>
+                    <a class="main-dellink" href="/?r=refereecomp/delete&rc=<?=$answer['referee'][$i]['rc']?>&comp=<?=$_GET['comp']?>">[X]</a>
+                </td>
+            <?}?>
+        </tr>
     <?}?>
-</div>
+    </tbody>
+</table>
